extract code into files
